% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rds_contents.R
\name{check.rds.contents}
\alias{check.rds.contents}
\title{Check RDS Contents}
\usage{
check.rds.contents(obj, required = list(), allow.extra = TRUE)
}
\arguments{
\item{obj}{The loaded RDS object.}

\item{required}{List of required elements.}

\item{allow.extra}{Logical, whether to allow additional, unexpected elements.
Default is \code{TRUE}.}
}
\value{
Returns errors if any found; otherwise, returns \code{TRUE}.
}
\description{
Checks the RDS file for expected elements.
}
