% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_parallel_lapply.R
\name{create.parallel.lapply}
\alias{create.parallel.lapply}
\title{Create Parallel Lapply}
\usage{
create.parallel.lapply(
  asp,
  exports,
  parallel = TRUE,
  threads = NULL,
  export.env = parent.frame(),
  dev.mode = FALSE,
  package.path = NULL,
  allow.mclapply.mac = FALSE
)
}
\arguments{
\item{asp}{The AutoSpectral parameter list.}

\item{exports}{The vector of variables and functions to pass to the clusters.}

\item{parallel}{Logical, controls whether parallel processing is used. Default
is \code{TRUE}.}

\item{threads}{Numeric, number of threads to use for parallel processing.
Default is \code{NULL} which will revert to \code{asp$worker.process.n} if
\code{parallel=TRUE}.}

\item{export.env}{The environment containing other functions and global
variables potentially needed by the clusters. Default is \code{parent.frame()}.}

\item{dev.mode}{Logical, allows testing of function while in development.
Default is \code{FALSE}.}

\item{package.path}{File.path to the R package files for AutoSpectral to
permit loading of the functions via \code{devtools::load_all()} while in \code{dev.mode}.
Default is \code{NULL}.}

\item{allow.mclapply.mac}{Logical, if \code{TRUE} permits \code{mclapply()} forking on
Mac OS. Default \code{FALSE} forces PSOCK cluster use on Mac to prevent
multithreaded Accelerate BLAS from crashing parallels when matrix ops are used.}
}
\value{
An lapply function, either based on \code{parLapply} for Windows or \code{mcLapply}
on Mac OS and Linux. If parallel backend initialization fails, sequential
\code{lapply} is returned.
}
\description{
Sets up parallel processing lapply function for Windows, Mac OS or Linux.
}
