% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_af.R
\name{fit.af}
\alias{fit.af}
\title{Fit Autofluorescence}
\usage{
fit.af(raw.data, unmixed, unmixing.matrix, spectra, af.spectra)
}
\arguments{
\item{raw.data}{Expression data from raw FCS files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{unmixed}{Unmixed data (unmixed using fluorophore spectra only). Cells
in rows and fluorophores in columns.}

\item{unmixing.matrix}{The OLS unmixing matrix for the fluorophore spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}.}
}
\value{
A list containing the adjusted unmixed data (\code{unmixed}), the abundances
of autofluorescence per cell (\code{AF}), the index corresponding to the spectrum
selected per cell (\code{af.idx}) and the raw space projection for the
autofluorescence (\code{fitted.af}).
}
\description{
Select and extract the best fitting autofluorescence spectrum per cell. Uses
minimization of fluorophore spillover as the criterion.
}
