% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fret.R
\name{fit.fret}
\alias{fit.fret}
\title{Fit FRET Signals}
\usage{
fit.fret(
  raw.data_subset,
  unmixed_subset,
  unmixing.matrix,
  spectra,
  flowcode.ids_subset,
  fret.spectra,
  flowcode.fluors,
  flowcode.combo.logical
)
}
\arguments{
\item{raw.data_subset}{Expression data from raw FCS files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra. Only provide the subset of data for the cells
that have FlowCodes.}

\item{unmixed_subset}{Unmixed data (unmixed using fluorophore spectra only).
Cells in rows and fluorophores in columns. Only provide the subset of data
for the cells that have FlowCodes.}

\item{unmixing.matrix}{The OLS unmixing matrix for the fluorophore spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{flowcode.ids_subset}{The FlowCode barcode IDs for each cell.}

\item{fret.spectra}{A list of matrices containing the FRET correction spectra
per barcode.}

\item{flowcode.fluors}{The names of the fluorophores used to identify the
FlowCodes}

\item{flowcode.combo.logical}{Integer (1/0) matrix describing the combinations
of fluorophores in each FlowCode barcode.}
}
\value{
A list containing the adjusted unmixed data (\code{unmixed}) and the raw
space projection for the FRET (\code{fitted.fret}).
}
\description{
Select and extract the best fitting FRET correction spectrum per cell. Uses
minimization of fluorophore spillover in the channels that do not belong to
that cell's FlowCode barcode as the criterion.
}
