% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcode_combo_plot.R
\name{flowcode.combo.plot}
\alias{flowcode.combo.plot}
\title{FlowCode Combo Plot}
\usage{
flowcode.combo.plot(
  unmixed,
  corrected,
  flowcode.fluors,
  asp,
  output.dir = "./flowcode_spectra",
  filename = "Corrected_biplots.jpg",
  width = 4,
  height = 4
)
}
\arguments{
\item{unmixed}{Matrix of unmixed flow cytometry expression data. Cells in
rows and fluorophores in columns. Fluorophores must include \code{flowcode.fluors}.}

\item{corrected}{Corrected matrix of unmixed flow cytometry expression data
after processing for FRET errors.}

\item{flowcode.fluors}{Named character vector of fluorophores used to identify
the FlowCodes. Names should be FlowCode epitope tags, values should be the
corresponding fluorophores.}

\item{asp}{The AutoSpectral parameter list. Prepare using
\code{get.autospectral.param}.}

\item{output.dir}{File path determining where the plot will be saved. Default
is \code{./flowcode_spectra}.}

\item{filename}{Character, name of the saved plot. Default is
\code{Corrected_biplots.jpg}.}

\item{width}{Numeric, default \code{5}. Width (in inches) of each panel.}

\item{height}{Numeric, default \code{5}. Height (in inches) of each panel.}
}
\value{
None. Saves the plot to the specified file path in \code{output.dir}.
}
\description{
Creates side-by-side pseudocolor density dot plots (standard flow cytometry)
for the FlowCode markers. Plots the unmixed data on a biexponential scale
using \code{AutoSpectral::create.biplot()}.
}
