% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_backend.R
\name{parallel.backend}
\alias{parallel.backend}
\title{Parallel Backend}
\usage{
parallel.backend(
  asp,
  exports,
  threads,
  export.env = parent.frame(),
  dev.mode = FALSE,
  package.path = NULL
)
}
\arguments{
\item{asp}{The AutoSpectral parameter list.}

\item{exports}{The vector of variables and functions to pass to the clusters.}

\item{threads}{Numeric, number of threads to use for parallel processing.}

\item{export.env}{The environment containing other functions and global
variables potentially needed by the clusters.}

\item{dev.mode}{Logical, allows testing of function while in development.
Default is \code{FALSE}.}

\item{package.path}{File.path to the R package files for AutoSpectral to
permit loading of the functions via \code{devtools::load_all()} while in \code{dev.mode}.}
}
\value{
An lapply function, either based on parLapply if parallel backend
initialization was successful or using sequential lapply if not.
}
\description{
Sets up parallel processing backend using parallel and parallelly to establish
PSOCK clusters on Windows. Sets up clusters, defines parLapply-based function.
}
