% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_expression_plot.R
\name{tag.expression.plot}
\alias{tag.expression.plot}
\title{Plot Tag Expression}
\usage{
tag.expression.plot(
  unmixed,
  flowcode.fluors,
  flowcode.ids,
  combo.df,
  asp,
  output.dir = "./flowcode_spectra",
  title = "Uncorrected",
  max.per.group = 2000,
  height = 20,
  width = 20
)
}
\arguments{
\item{unmixed}{Matrix of unmixed flow cytometry expression data. Cells in
rows and fluorophores in columns. Fluorophores must include \code{flowcode.fluors}.}

\item{flowcode.fluors}{Named character vector of fluorophores used to identify
the FlowCodes. Names should be FlowCode epitope tags, values should be the
corresponding fluorophores.}

\item{flowcode.ids}{Numeric vector containing the FlowCode combination
assignments per cell (result of \code{debarcode()}).}

\item{combo.df}{Data frame describing the valid combinations of FlowCodes.
Structure: One row per combination. Columns are \code{Id}, \code{Procode.tag1},
\code{Procode.tag2} and \code{Procode.tag3}, describing the name (e.g., CRISPR target),
and three epitopes for the combination, respectively.}

\item{asp}{The AutoSpectral parameter list.}

\item{output.dir}{File path determining where the plot will be saved. Default
is \code{./flowcode_spectra}.}

\item{title}{Character, name to add to the saved plots. Default is \code{Uncorrected}.}

\item{max.per.group}{Numeric, maximum number of cells to plot per group
(FlowCode combination). Larger numbers will take longer. Default is \code{2000}.}

\item{height}{Numeric, default \code{20}. Width (in inches) of the saved plot.}

\item{width}{Numeric, default \code{20}. Height (in inches) of the saved plot.}
}
\value{
None. Saves the plot to the specified file path in \code{output.dir}.
}
\description{
Creates ridgeline (density) plots of the FlowCode fluorophore expression
levels from unmixed data.
}
