% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_flowcode_folder.R
\name{unmix.flowcode.folder}
\alias{unmix.flowcode.folder}
\title{Unmix All FCS Files in a Directory}
\usage{
unmix.flowcode.folder(
  fcs.dir,
  spectra,
  asp,
  flow.control,
  af.spectra,
  spectra.variants,
  flowcode.spectra,
  thresholds.file = NULL,
  k = 10,
  output.dir = NULL,
  file.suffix = NULL,
  include.imaging = FALSE,
  parallel = TRUE,
  threads = if (parallel) 0 else 1,
  verbose = TRUE,
  optimize = TRUE
)
}
\arguments{
\item{fcs.dir}{Directory containing FCS files to be unmixed.}

\item{spectra}{Matrix containing spectra information.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}. Required for \code{FlowCodeUnmixe} unmixing.}

\item{spectra.variants}{Named list (names are fluorophores) carrying matrices
of spectral signature variations for each fluorophore. Prepare using
\code{get.spectral.variants}.}

\item{flowcode.spectra}{Structured output from \code{get.flowcode.spectra()}, which
details the combination-level spectral unmixing errors due to FRET-like
artefacts.}

\item{thresholds.file}{Optional, file name and path to the thresholds CSV file
produced using the threshold setting app. Thresholds are provided by default
as part of \code{flowcode.spectra}, but those will have been selected on the
FlowCode backbone sample. If the thresholds for positivity of any of the
fluorophores used to identify the FlowCodes differs in the fully stained
samples being unmixed here, provide a new set of thresholds from the
ThresholdApp. Run this on an unmixed sample (OLS from the instrument is fine),
and refer to the new thresholds file here. Default is \code{NULL}, which will be
ignored.}

\item{k}{Numeric, controls the number of variants tested for each fluorophore,
autofluorescence and FRET spectrum. Default is \code{10}. Values up to \code{10} provide
additional benefit in unmixing quality, \code{1} will be fastest. Only used if
\code{optimize=TRUE}.}

\item{output.dir}{Directory to save the unmixed FCS files
(default is \code{asp$unmixed.fcs.dir}).}

\item{file.suffix}{A character string to append to the output file name.
Default is \code{NULL}}

\item{include.imaging}{Logical indicating whether to include imaging data in
the written FCS file: relevant for S8 and A8. Default is \code{FALSE}}

\item{parallel}{Logical, default is \code{TRUE}, which enables parallel processing
for per-cell unmixing methods.}

\item{threads}{Numeric, defaults to all available cores if \code{parallel=TRUE}.}

\item{verbose}{Logical, controls messaging. Default is \code{TRUE}.}

\item{optimize}{Logical, whether to perform per-cell spectral optimization.
Faster without this, usually better with it.}
}
\value{
None. Saves the unmixed FCS files to the specified output directory.
}
\description{
This function unmixes all FCS files in a specified directory using the
provided spectra and method, and saves the unmixed FCS files to an output
directory of the user's choice.
}
