% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flowcode_spectra.R
\name{get.flowcode.spectra}
\alias{get.flowcode.spectra}
\title{Get FlowCode Spectra}
\usage{
get.flowcode.spectra(
  backbone.rds,
  thresholds.file,
  asp,
  output.dir = "./flowcode_spectra",
  filename = "FlowCode_Spectra.rds",
  figures = TRUE,
  plot.corrections = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{backbone.rds}{File name and path to the full (large) rds file produced
by running \code{unmix.backbone}. For this to work well, all combinations present
in the experimental samples should be well represented in the
backbone sample used.}

\item{thresholds.file}{File name and path to the thresholds CSV file produced
using the threshold setting app.}

\item{asp}{The AutoSpectral parameter list. Prepare using
\code{get.autospectral.param}.}

\item{output.dir}{Destination folder for output plots and saved spectra.
Default is \code{./flowcode_spectra}.}

\item{filename}{Logical, whether to create and save plots of the data
correction. Default is \code{TRUE}.}

\item{figures}{Logical, controls whether graphs of the process are produced.
Default is \code{TRUE}.}

\item{plot.corrections}{Logical, default is \code{FALSE}. If \code{TRUE}, the FRET
unmixing will be applied to \code{flowcode.backbone.fcs} and plots of the
corrected data (FlowCode channels only) will be produced in \code{output.dir}.}

\item{verbose}{Logical, default is \code{TRUE}. Set to \code{FALSE} to suppress
messaging.}
}
\value{
A named list with two elements:
\enumerate{
\item Thresholds: numeric thresholds for debarcoding
\item FRET: A list of matrices of FRET spectral variations per barcode combo
\item Flowcode.fluors: Character vector of fluorophores linked to the FlowCodes
\item Combos: The valid combinations, specified by the combination file.
}
}
\description{
Measure spectral unmixing errors per barcode in a sample stained only with
FlowCode tag antibodies (\code{flowcode.backbone.fcs}). Measures FRET error and
returns the calculated correction matrices and thresholds for debarcoding.
}
