% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_threshold_app.R
\name{launch.threshold.app}
\alias{launch.threshold.app}
\title{Launch the Thresholding Shiny App}
\usage{
launch.threshold.app(
  flowcode.combo.file = NULL,
  flow.control = NULL,
  spectra = NULL
)
}
\arguments{
\item{flowcode.combo.file}{Optional, default is \code{NULL}. Provide this when you
wish to establish new thresholds using the app and will be working with an
FCS file rather than loading the RDS object. When given: File name and path
to the CSV file containing the information describing your FlowCode library.
Describes the valid combinations of FlowCodes. Structure: One row per
combination. Columns are \code{Id}, \code{Procode.tag1}, \code{Procode.tag2} and
\code{Procode.tag3}, describing the name (e.g., CRISPR target), and three epitopes
for the combination, respectively.}

\item{flow.control}{Optional, default is \code{NULL}. Provide this when you
wish to establish new thresholds using the app and will be working with an
FCS file rather than loading the RDS object. A list containing flow cytometry
control parameters.}

\item{spectra}{Optional, default is \code{NULL}. Provide this when you wish to
establish new thresholds using the app and will be working with an FCS file
rather than loading the RDS object.Spectral signatures of fluorophores,
normalized between 0 and 1, with fluorophores in rows and detectors in columns.}
}
\description{
A simple wrapper function to launch the manual threshold selection app.
}
