% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_keywords.R
\name{prep.keywords}
\alias{prep.keywords}
\title{Prepare Keywords}
\usage{
prep.keywords(
  fcs.keywords,
  final.matrix,
  original.param,
  spectra,
  af.spectra,
  flow.control,
  asp,
  method,
  file.name,
  combos,
  include.imaging = FALSE
)
}
\arguments{
\item{fcs.keywords}{The input keywords obtained from the read FCS file.}

\item{final.matrix}{The expression data, containing both unmixed data and any
retained parameters such as scatter and time.}

\item{original.param}{The original parameter (column) names of the input FCS
expression data.}

\item{spectra}{A matrix containing the spectral data.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns.}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{asp}{The AutoSpectral parameter list.}

\item{method}{A character string specifying the unmixing method used.}

\item{file.name}{The name of the FCS file to be written.}

\item{combos}{A character vector of the FlowCode combination tag names.}
}
\value{
The updated keyword list for writing the FCS file.
}
\description{
Updates FCS file keywords after unmixing to define the new parameters. Tracks
existing keywords from the input FCS file for metadata compatibility.
}
