% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFCS.R
\name{readFCS}
\alias{readFCS}
\title{Read FCS File}
\usage{
readFCS(fcs.path, keywords = FALSE)
}
\arguments{
\item{fcs.path}{A character string specifying the file path (directory and
file name) for the .fcs file to be read.}

\item{keywords}{Logical, default \code{FALSE}. Controls whether keywords are
returned as well as the expression data matrix (useful for writing a new file).}
}
\value{
If \code{keywords = TRUE}, a list containing two elements:
\enumerate{
\item the expression data in a matrix, and 2) the keywords.
If \code{keywords = FALSE}, only the expression data is returned as a matrix.
}
}
\description{
A light weight FCS file binary reader based on \code{flowstate}. Lower memory usage
and faster than \code{flowCore::read.FCS()}.
}
