% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_backbone.R
\name{unmix.backbone}
\alias{unmix.backbone}
\title{Unmix Backbone}
\usage{
unmix.backbone(
  flowcode.backbone.fcs,
  spectra,
  af.spectra,
  flow.control,
  asp,
  flowcode.combo.file,
  n.cells.gate = 100,
  output.dir = "./flowcode_spectra",
  filename = "FlowCode_Backbone.rds",
  verbose = TRUE
)
}
\arguments{
\item{flowcode.backbone.fcs}{File name and path for the FlowCode backbone
control FCS file. This should be a sample of cells, ideally the same cell
source as your single-stained control samples. These cells should be stained
with all the FlowCode epitope tag antibodies present in the fully stained
sample and nothing else. All tag combinations should be present and well
represented for best results. A minimum of 100 cells per combination is
recommended; 2000+ is ideal.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}.}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{asp}{The AutoSpectral parameter list. Prepare using
\code{get.autospectral.param}.}

\item{flowcode.combo.file}{File name and path to the CSV file containing the
information describing your FlowCode library. Describes the valid combinations
of FlowCodes. Structure: One row per combination. Columns are \code{Id},
\code{Procode.tag1}, \code{Procode.tag2} and \code{Procode.tag3}, describing the name (e.g.,
CRISPR target), and three epitopes for the combination, respectively.}

\item{n.cells.gate}{Numeric, default \code{100}. The number of cells that will be
used to define the gate. A gate region will be defined using \code{define.gate()}
around the brightest \code{n.cells.gate} for each fluorophore used to identify
FlowCode epitope tags. Scatter coordinates of all brightest events will be
combined to determine the optimal gating region.}

\item{output.dir}{Destination folder for output plots and saved spectra.
Default is \code{./flowcode_spectra}.}

\item{filename}{Name of the output RDS file, default is \code{FlowCode_Backbone.rds}.}

\item{verbose}{Logical, whether to send messages to the console.
Default is \code{TRUE}.}
}
\value{
None. Saves RDS objects containing the processed data to disk.
}
\description{
Unmix the backbone sample, stained only with anti-FlowCode epitope antibodies,
for downstream identification of FRET/unmixing errors on a per-combo basis.
}
