% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_flowcode_fcs.R
\name{unmix.flowcode.fcs}
\alias{unmix.flowcode.fcs}
\title{Unmix FlowCode FCS Data}
\usage{
unmix.flowcode.fcs(
  fcs.file,
  spectra,
  asp,
  flow.control,
  af.spectra,
  spectra.variants,
  flowcode.spectra,
  thresholds.file = NULL,
  weights = NULL,
  cell.weighting = FALSE,
  cell.weight.regularize = TRUE,
  k = 10,
  output.dir = NULL,
  file.suffix = NULL,
  include.imaging = FALSE,
  parallel = TRUE,
  threads = if (parallel) 0 else 1,
  verbose = TRUE
)
}
\arguments{
\item{fcs.file}{A character string specifying the path to the FCS file.}

\item{spectra}{A matrix containing the spectral data.}

\item{asp}{The AutoSpectral parameter list.
Prepare using \code{get.autospectral.param}}

\item{flow.control}{A list containing flow cytometry control parameters.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}. Required for \code{AutoSpectral} unmixing. Default is
\code{NULL} and will thus provoke failure if no spectra are provided and
\code{AutoSpectral} is selected.}

\item{spectra.variants}{Named list (names are fluorophores) carrying matrices
of spectral signature variations for each fluorophore. Prepare using
\code{get.spectral.variants}.}

\item{flowcode.spectra}{Structured output from \code{get.flowcode.spectra()}, which
details the combination-level spectral unmixing errors due to FRET-like
artefacts.}

\item{thresholds.file}{Optional, file name and path to the thresholds CSV file
produced using the threshold setting app. Thresholds are provided by default
as part of \code{flowcode.spectra}, but those will have been selected on the
FlowCode backbone sample. If the thresholds for positivity of any of the
fluorophores used to identify the FlowCodes differs in the fully stained
samples being unmixed here, provide a new set of thresholds from the
ThresholdApp. Run this on an unmixed sample (OLS from the instrument is fine),
and refer to the new thresholds file here. Default is \code{NULL}, which will be
ignored.}

\item{weights}{Optional numeric vector of weights (one per fluorescent
detector). Default is \code{NULL}, in which case weighting will be done by
channel means (Poisson variance).}

\item{cell.weighting}{Logical, whether to use cell-specific weighting for a
more Poisson-like unmixing. Default is \code{FALSE}.}

\item{cell.weight.regularize}{Logical, whether to regularize cell-specific
weights towards the bulk mean weighting set by \code{weights}. 50:50 averaging.
Default is \code{TRUE}. Only active if \code{cell.weighting=TRUE}.}

\item{k}{Numeric, controls the number of variants tested for each fluorophore,
autofluorescence and FRET spectrum. Default is \code{10}. Values up to \code{10} provide
additional benefit in unmixing quality, \code{1} will be fastest.}

\item{output.dir}{A character string specifying the directory to save the
unmixed FCS file. Default is \code{NULL}, which reverts to \code{asp$unmixed.fcs.dir}.}

\item{file.suffix}{A character string to append to the output file name.
Default is \code{NULL}.}

\item{include.imaging}{A logical value indicating whether to include imaging
parameters in the written FCS file. Default is \code{FALSE}.}

\item{parallel}{Logical, default is \code{TRUE}, which enables parallel processing
for per-cell unmixing methods.}

\item{threads}{Numeric, default is \code{NULL}, in which case \code{asp$worker.process.n}
will be used. \code{asp$worker.process.n} is set by default to be one less than the
available cores on the machine. Multi-threading is only used if \code{parallel} is
\code{TRUE}.}

\item{verbose}{Logical, whether to send messages to the console.
Default is \code{TRUE}.}
}
\value{
None. The function writes the unmixed FCS data to a file.
}
\description{
This function performs spectral unmixing on FCS data from FlowCode samples.
}
