% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_flowcode.R
\name{unmix.flowcode}
\alias{unmix.flowcode}
\title{Unmix FlowCode}
\usage{
unmix.flowcode(
  raw.data,
  spectra,
  af.spectra,
  spectra.variants,
  flowcode.spectra,
  asp,
  thresholds = NULL,
  k = 1,
  parallel = TRUE,
  threads = if (parallel) 0 else 1,
  verbose = TRUE,
  optimize = TRUE
)
}
\arguments{
\item{raw.data}{Expression data from raw FCS files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}.}

\item{spectra.variants}{Named list (names are fluorophores) carrying matrices
of spectral signature variations for each fluorophore. Prepare using
\code{get.spectral.variants()}. Default is \code{NULL}.}

\item{flowcode.spectra}{Structured output from \code{get.flowcode.spectra()}, which
details the combination-level spectral unmixing errors due to FRET-like
artefacts.}

\item{asp}{The AutoSpectral parameter list. Prepare using
\code{get.autospectral.param}.}

\item{thresholds}{Optional named numeric vector of positivity thresholds for
the FlowCode fluorophores. Overrides the thresholds provided by \code{flowcode.spectra}.
Default is \code{NULL}, which is unused.}

\item{k}{Numeric, controls the number of variants tested for each fluorophore,
autofluorescence and FRET spectrum. Default is \code{10}, which will be good, \code{1}
is fastest. Values up to \code{10} provide additional benefit in unmixing quality.
Only used if \code{optimize=TRUE}.}

\item{parallel}{Logical, whether to use parallel processing for the per-cell
unmixing. Default is \code{FALSE}.}

\item{threads}{Numeric. Number of threads to use for parallel processing.
Defaults to \code{1} for sequential processing, or \code{0} (all cores) if \code{parallel=TRUE}.}

\item{verbose}{Logical, whether to send messages to the console.
Default is \code{TRUE}.}

\item{optimize}{Logical, whether to perform per-cell spectral optimization.
Faster without this, usually better with it.}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Unmix FlowCode samples, correcting FRET errors and debarcoding the data.
}
